requirements.txt
aiofiles==23.2.1
aiogram==2.15
aiohttp==3.7.4
aioredis==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-property==0.2.2
async-timeout==3.0.1
asyncpg==0.28.0
asyncstdlib==3.10.9
attrs==23.1.0
Babel==2.13.1
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
curl-cffi==0.5.10
duckduckgo-search==3.9.11
exceptiongroup==1.2.0
fastapi==0.104.1
Flask==3.0.0
g4f==0.1.9.2
gino==1.0.1
h11==0.14.0
h2==4.1.0
hiredis==2.2.3
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
Js2Py==0.74
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
mmh3==4.0.1
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
outcome==1.3.0.post0
Pillow==10.1.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
py-arkose-generator==0.0.0.2
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyExecJS==1.5.1
pyjsparser==2.7.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.24
starlette==0.27.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzlocal==5.2
undetected-chromedriver==3.5.4
urllib3==2.1.0
uvicorn==0.24.0.post1
waitress==2.1.2
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
